Add custom segment generator appliance
The segment generator provided by TV Kitchen only watches stream
container payloads; since we aren't working with a stream of video (only
a stream of text) we have to write our own for now.

We need segments to be generated because that's how we'll know when to
create new files / upload a file set

I'll be opening a PR to improve the TV Kitchen version so that it can
take other types of "heartbeat" payloads in future, at which point we'll
want to switch over to that one.

Issue #7